+++ cmd/telemetry-gateway/internal/server/BUILD.bazel
@@ -4,7 +4,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "server",
     srcs = [
-        "managed_service.go",
         "metrics.go",
         "migrate_events.go",
         "publish_events.go",
@@ -14,7 +13,7 @@ go_library(
     visibility = ["//cmd/telemetry-gateway:__subpackages__"],
     deps = [
         "//cmd/telemetry-gateway/internal/events",
-        "//internal/authbearer",
+        "//cmd/telemetry-gateway/internal/server/samsm2m",
         "//internal/licensing",
         "//internal/pubsub",
         "//internal/sams",
@@ -26,7 +25,6 @@ go_library(
         "@io_opentelemetry_go_otel//attribute",
         "@io_opentelemetry_go_otel_metric//:metric",
         "@org_golang_google_grpc//codes",
-        "@org_golang_google_grpc//metadata",
         "@org_golang_google_grpc//status",
     ],
 )
